pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mentat-lss"
description = "package for emulating galaxy power spectrum multipoles using neural networks"
version = "0.9.3"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Joe Adamo", email = "jadamo@arizona.edu" }
]
requires-python = ">= 3.9"
dependencies = [
    "setuptools",
    "build",
    "numpy<2",
    "scipy",
    "torch",
    "pytest",
    "tqdm",
    "pyyaml",
    "symbolic_pofk@git+https://github.com/DeaglanBartlett/symbolic_pofk.git"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://github.com/jadamo/mentat-lss"
